broken timeslider displaying
